........S. [ZBXNEXT-3926] changed DB_TRIGGER structure member value type to unsigned char

+++ src/zabbix_server/events.c
@@ -128,7 +128,7 @@ int	add_event(unsigned char source, unsigned char object, zbx_uint64_t objectid,
 		events[events_num].trigger.type = trigger_type;
 		events[events_num].trigger.correlation_mode = trigger_correlation_mode;
 		events[events_num].trigger.correlation_tag = zbx_strdup(NULL, trigger_correlation_tag);
-		events[events_num].trigger.value = (int)trigger_value;
+		events[events_num].trigger.value = trigger_value;
 
 		substitute_simple_macros(NULL, &events[events_num], NULL, NULL, NULL, NULL, NULL, NULL,
 				&events[events_num].trigger.correlation_tag, MACRO_TYPE_TRIGGER_TAG, NULL, 0);
